Password Grant Flow and Nonrelational Support
Implemented the password grant type. Returns an access token given a client id and a username and password.
Added a column called trusted on the Client model. Only trusted clients can use the password grant flow.

Replaced all of the scope ManyToMany fields with generic relations to the new ScopeMap model that handles the ManyToMany relations in a nonrel friendly way.